deploy/olm-catalog/requirements.txt
python-dxf
ruamel.yaml
